[Refactor] 스프링 시큐리티 적용

+++ src/main/java/com/study/toDoList/service/TaskService.java
@@ -8,6 +8,7 @@ import com.study.toDoList.dto.TaskSaveDto;
 import com.study.toDoList.dto.TaskUpdateDto;
 import com.study.toDoList.exception.ex.MyErrorCode;
 import com.study.toDoList.exception.ex.MyNotFoundException;
+import com.study.toDoList.exception.ex.MyNotPermittedException;
 import com.study.toDoList.repositoy.MemberRepository;
 import com.study.toDoList.repositoy.TaskRepository;
 import lombok.RequiredArgsConstructor;
@@ -31,15 +32,21 @@ public class TaskService {
     }
 
     @Transactional
-    public Long update(Long id, TaskUpdateDto taskUpdateDto){
+    public Long update(Long id, TaskUpdateDto taskUpdateDto, Long loginId){
         Task task =taskRepository.findById(id).orElseThrow(()-> new MyNotFoundException(MyErrorCode.TASK_NOT_FOUND));
+        if(task.getMember().getId()!=loginId){
+            throw new MyNotPermittedException(MyErrorCode.UPDATE_NOT_PERMITTED);
+        }
         task.update(taskUpdateDto.getTitle(),taskUpdateDto.getDescription(),taskUpdateDto.getEndDate());
         return id;
     }
 
     @Transactional
-    public void delete(Long id){
+    public void delete(Long id,Long loginId){
         Task task =taskRepository.findById(id).orElseThrow(()-> new MyNotFoundException(MyErrorCode.TASK_NOT_FOUND));
+        if(task.getMember().getId()!=loginId){
+            throw new MyNotPermittedException(MyErrorCode.DELETE_NOT_PERMITTED);
+        }
         taskRepository.delete(task);
     }
 
@@ -56,8 +63,11 @@ public class TaskService {
     }
 
     @Transactional
-    public TaskResponseDto changeIsFinished(Long id){
+    public TaskResponseDto changeIsFinished(Long id, Long loginId){
         Task task =taskRepository.findById(id).orElseThrow(()-> new MyNotFoundException(MyErrorCode.TASK_NOT_FOUND));
+        if(task.getMember().getId()!=loginId){
+            throw new MyNotPermittedException(MyErrorCode.DELETE_NOT_PERMITTED);
+        }
         task.changeIsFinished();
         return new TaskResponseDto(task);
     }
